Small refactoring + Add example fetch + Fake Viewer Data
