+++ newsfeed/server/printSchema.mjs
@@ -1,5 +1,5 @@
 import {printSchema} from 'graphql';
 
-import {schema} from './graphql.mjs';
+import {schema} from './schema.mjs';
 
-console.log("# @generated\n\n" + printSchema(schema));
+console.log("# @generated\n# run `yarn print-schema` to re-generate this file.\n\n" + printSchema(schema));
