+++ newsfeed/server/rootValue.mjs
@@ -1,0 +1,16 @@
+
+export const rootValue = {
+  viewer: () => {
+    return {
+      actor: {
+        __typename: 'User',
+        id: '1',
+        name: 'Alice',
+      }
+    };
+  },
+  node: () => {
+    // TODO: implement
+    return null
+  },
+};
